+++ poetry/repositories/pool.py
@@ -62,13 +62,13 @@ class Pool(BaseRepository):
     def has_package(self, package):
         raise NotImplementedError()
 
-    def package(self, name, version):
+    def package(self, name, version, extras=None):
         package = poetry.packages.Package(name, version, version)
         if package in self._packages:
             return self._packages[self._packages.index(package)]
 
         for repository in self._repositories:
-            package = repository.package(name, version)
+            package = repository.package(name, version, extras=extras)
             if package:
                 self._packages.append(package)